Função de inserir(revisada), re-hash e coerência de inicialização
- Função de inserir agora atualiza registro se palavra já está no dicionário
-Consertado bugs na função de re-hash(limpava completamente o dicionário)
-Coerência de inicialização de worddata ao criar worddata de palavra extraida do tweet, usado nas demais funções.

+++ core/dictionary.h
@@ -6,12 +6,12 @@
 //Stores the information about the given word
 typedef struct wordData {
     utf8_string word;
-    int weight;
+    float weight;
     int occurrences; //Number of occurences
     int score;  //Accumulated score
 
     wordData() :
-    word(""), weight(0), occurrences(0), score(0) { }
+    word(""), weight(0.0), occurrences(0), score(0) { }
 
 } wordData;
